// source/apps/system/auth/api/users/schemas/users_list.output.ts
/**
 * This file was automatically generated by source/platform/system/schemas/json2ts.ts
 * DO NOT MODIFY IT MANUALLY. Instead, modify the source JSON Schema file,
 * and run 'npm run schemas' to regenerate this file.
 * The source file has the same name as this one but has the extension .json
 */

export interface OutputSystemAuthUsersList {
  count: number;
  rows: {
    id: string;
    username: string;
    fullname: string;
    email: string;
    timezone?: string;
    country?: string;
    language?: string;
    country_text?: string;
    language_text?: string;
    enabled?: string;
  }[];
}