// source/apps/system/auth/api/users/schemas/user_edit.input.ts
/**
 * This file was automatically generated by source/platform/system/schemas/json2ts.ts
 * DO NOT MODIFY IT MANUALLY. Instead, modify the source JSON Schema file,
 * and run 'npm run schemas' to regenerate this file.
 * The source file has the same name as this one but has the extension .json
 */

export interface InputSystemAuthUserEdit {
  id: string;
  username?: string;
  fullname?: string;
  email?: string;
  password?: string;
  language?: string;
  country?: string;
  timezone?: string;
  enabled?: string;
}